refactor: header method to headers.go
